login required and some minor fixes

+++ base/views.py
@@ -1,4 +1,5 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.decorators import login_required
@@ -8,6 +9,8 @@ from .models import Room, Topic
 from .forms import RoomForm
 
 def loginPage(request):
+  if request.user.is_authenticated:
+    return redirect('home')
   if request.method == 'POST':
     username = request.POST.get('username')
     password = request.POST.get('password')
@@ -51,9 +54,13 @@ def createRoom(request):
   context={ "form": form }
   return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='/login')
 def updateRoom(request, pk):
   room = Room.objects.get(pk=pk)
   form = RoomForm(instance=room)
+  if request.user != room.host:
+    return HttpResponse('You are not allowed here!')
+
   if request.method == 'POST':
     form = RoomForm(request.POST, instance=room)
     if form.is_valid:
@@ -63,8 +70,11 @@ def updateRoom(request, pk):
   context={ "form": form }
   return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
   room = Room.objects.get(pk=pk)
+  if request.user != room.host:
+    return HttpResponse('You are not allowed here!')
   if request.method == 'POST':
     room.delete()
     return redirect('home')
